PR #10 review changes
- Removed unnecessary config from webpack config

- Remove html template

- Removed displayNames from tables, changed for a title inside

- Updated README

- Moved contents outside of styled-components folder, since the benchmarks are not all related to styled-components. Everything is in a single directory now.
