+++ benchmarks/README.md
@@ -1,0 +1,21 @@
+# xstyled Benchmarks
+
+A collection of `xstyled` benchmarks
+
+## Web
+
+Use this to benchmark `@xstyled/styled-components` against other libraries.
+
+```bash
+npm start
+```
+
+The app will run on `http://localhost:8083`
+
+## CLI
+
+Use this to benchmark `@xstyled/system` against other libraries
+
+```bash
+npm run benchmark
+```
