+++ benchmarks/src/components/BigTable.js
@@ -97,7 +97,6 @@ export default function App({ Table }) {
   return (
     <div>
       <h2>The higher the percentage, the faster was the cell render</h2>
-      <h3>{Table.displayName}</h3>
       <div>
         <input type="submit" value="Generate" onClick={handleClick} />{' '}
         <span>{getUniqueSize(table)} unique cells</span>
