+++ benchmarks/src/index.js
@@ -2,6 +2,7 @@ import React from 'react'
 import { render } from 'react-dom'
 import App from './App'
 
-const root = document.querySelector('#root')
+const root = document.createElement('div')
+document.body.appendChild(root)
 
 render(<App />, root)
