--- benchmarks/styled-components/README.md
@@ -1,21 +1,0 @@
-# styled-components Benchmark
-
-Use this to benchmark `@xstyled/styled-components` against `styled-components`.
-
-## Web
-
-For a React render performance test, run
-
-```bash
-npm start
-```
-
-The app will run on `http://localhost:8083`
-
-## CLI
-
-To benchmark `@xstyled/system` against `styled-system` run
-
-```bash
-npm run benchmark
-```
