+++ benchmarks/webpack.config.js
@@ -6,11 +6,6 @@ const SRC_DIR = path.resolve(__dirname, 'src')
 module.exports = {
   mode: 'production',
   entry: path.resolve(SRC_DIR, 'index.js'),
-  output: {
-    path: `${__dirname}/public`,
-    filename: 'bundle.js',
-    publicPath: '/',
-  },
   devServer: {
     inline:true,
     contentBase: './public',
@@ -26,8 +21,6 @@ module.exports = {
     ],
   },
   plugins: [new HtmlWebpackPlugin({
-    template: `${__dirname}/src/index.html`,
-    filename: 'index.html',
-    inject: 'body',
+    title: 'xstyled Benchmark',
   })],
 }
